requirements.txt
anthropic>=0.40.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=7.0.0
